requirements.txt
django==4.0.4
djangorestframework
psycopg2-binary
django-cors-headers
django-environ
Pillow
djangorestframework-simplejwt
django-filter